backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
moviepy==1.0.3
pydub==0.25.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
httpx==0.25.2

